[#MHV-38997] Refactored new folder modal
